Factor out the functionality of best config tracking into a separate abstract Optimizer class (#709)
