+++ mlos_bench/mlos_bench/optimizers/mlos_core_optimizer.py
@@ -169,7 +169,6 @@ class MlosCoreOptimizer(Optimizer):
             df_config = self._to_df([tunables.get_param_values()])
             _LOG.debug("Score: %s Dataframe:\n%s", score, df_config)
             self._opt.register(df_config, pd.Series([score], dtype=float))
-        self._iter += 1
         return score
 
     def get_best_observation(self) -> Union[Tuple[float, TunableGroups], Tuple[None, None]]:
