+++ mlos_bench/mlos_bench/optimizers/mock_optimizer.py
@@ -9,19 +9,19 @@ Mock optimizer for mlos_bench.
 import random
 import logging
 
-from typing import Callable, Dict, Optional, Sequence, Tuple, Union
+from typing import Callable, Dict, Optional, Sequence
 
 from mlos_bench.environments.status import Status
 from mlos_bench.tunables.tunable import Tunable, TunableValue
 from mlos_bench.tunables.tunable_groups import TunableGroups
 
-from mlos_bench.optimizers.base_optimizer import Optimizer
+from mlos_bench.optimizers.track_best_optimizer import TrackBestOptimizer
 from mlos_bench.services.base_service import Service
 
 _LOG = logging.getLogger(__name__)
 
 
-class MockOptimizer(Optimizer):
+class MockOptimizer(TrackBestOptimizer):
     """
     Mock optimizer to test the Environment API.
     """
@@ -38,8 +38,6 @@ class MockOptimizer(Optimizer):
             "float": lambda tunable: rnd.uniform(*tunable.range),
             "int": lambda tunable: rnd.randint(*tunable.range),
         }
-        self._best_config: Optional[TunableGroups] = None
-        self._best_score: Optional[float] = None
 
     def bulk_register(self, configs: Sequence[dict], scores: Sequence[Optional[float]],
                       status: Optional[Sequence[Status]] = None, is_warm_up: bool = False) -> bool:
@@ -71,20 +69,3 @@ class MockOptimizer(Optimizer):
                 tunable.value = self._random[tunable.type](tunable)
         _LOG.info("Iteration %d :: Suggest: %s", self._iter, tunables)
         return tunables
-
-    def register(self, tunables: TunableGroups, status: Status,
-                 score: Optional[Union[float, dict]] = None) -> Optional[float]:
-        registered_score = super().register(tunables, status, score)
-        if status.is_succeeded() and (
-            self._best_score is None or (registered_score is not None and registered_score < self._best_score)
-        ):
-            self._best_score = registered_score
-            self._best_config = tunables.copy()
-        self._iter += 1
-        return registered_score
-
-    def get_best_observation(self) -> Union[Tuple[float, TunableGroups], Tuple[None, None]]:
-        if self._best_score is None:
-            return (None, None)
-        assert self._best_config is not None
-        return (self._best_score * self._opt_sign, self._best_config)
